name refactoring & useless models
