+++ api/anri/apps/products/models/employee.py
@@ -1,10 +1,10 @@
 from django.db import models
 
-from anri.apps.products.models.users import User
+from anri.apps.products.models.user import User
 from anri.apps.products.models.contragents import Contragent
 from anri.apps.common.models import CoreModel
 
 
 class Employee(CoreModel):
-    user_id = models.ForeignKey(User)
+    user_id = models.ForeignKey(User.uuid)
     contragent_id = models.ForeignKey(Contragent)
